Folder structure for project
